+++ game/core/src/abilities/shatter_armor_ability.ts
@@ -9,9 +9,8 @@
  * -----------------------------------------------------------------------------
  */
 
-import { AbilityType, HoCScene } from "@heroesofcrypto/common";
+import { AbilityType, HoCScene, Unit } from "@heroesofcrypto/common";
 
-import { Unit } from "../units/units";
 import { getLapString } from "../utils/strings";
 
 export function processShatterArmorAbility(
