+++ game/core/src/abilities/skewer_strike_ability.ts
@@ -9,11 +9,19 @@
  * -----------------------------------------------------------------------------
  */
 
-import { AttackType, Grid, GridSettings, GridMath, HoCMath, HoCConstants, HoCScene } from "@heroesofcrypto/common";
+import {
+    AttackType,
+    Grid,
+    GridSettings,
+    GridMath,
+    HoCMath,
+    HoCConstants,
+    HoCScene,
+    Unit,
+} from "@heroesofcrypto/common";
 
 import { FightStateManager } from "../state/fight_state_manager";
 import { DamageStatisticHolder } from "../stats/damage_stats";
-import { Unit } from "../units/units";
 import { UnitsHolder } from "../units/units_holder";
 import { nextStandingTargets } from "./abilities_helper";
 import { processAggrAbility } from "./aggr_ability";
