+++ game/core/src/abilities/through_shot_ability.ts
@@ -9,12 +9,11 @@
  * -----------------------------------------------------------------------------
  */
 
-import { AttackType, HoCLib, HoCMath, HoCConstants, Grid, HoCScene } from "@heroesofcrypto/common";
+import { AttackType, HoCLib, HoCMath, HoCConstants, Grid, HoCScene, Unit } from "@heroesofcrypto/common";
 
 import { Drawer } from "../draw/drawer";
 import { FightStateManager } from "../state/fight_state_manager";
 import { DamageStatisticHolder } from "../stats/damage_stats";
-import { Unit } from "../units/units";
 import { UnitsHolder } from "../units/units_holder";
 import { processLuckyStrikeAbility } from "./lucky_strike_ability";
 import { processPetrifyingGazeAbility } from "./petrifying_gaze_ability";
