+++ game/core/src/units/units_holder.ts
@@ -21,12 +21,12 @@ import {
     EffectHelper,
     HoCLib,
     SquarePlacement,
+    Unit,
     HoCConstants,
 } from "@heroesofcrypto/common";
 
 import { FightStateManager } from "../state/fight_state_manager";
 import { RenderableUnit } from "./renderable_unit";
-import { Unit } from "./units";
 
 export class UnitsHolder {
     private readonly grid: Grid;
@@ -266,6 +266,9 @@ export class UnitsHolder {
                 const unitsUpper: Unit[] = [];
                 const unitsLower: Unit[] = [];
                 for (const u of this.getAllUnitsIterator()) {
+                    if (u.isDead()) {
+                        continue;
+                    }
                     if (u.getTeam() === TeamType.LOWER) {
                         unitsLower.push(u);
                     } else {
